rename control_methods to control_strategies
